requirements.txt
numpy
pandas
pyyaml
cornac
pandas
requests
matplotlib
scikit-learn